added firebase for games; linted

+++ app/src/main/java/com/dav1337d/catalog/db/RoomBook.kt
@@ -8,18 +8,18 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "roomBook", indices = arrayOf(Index(value = ["title"], unique = true)))
 data class RoomBook(
     @PrimaryKey val uid: Int? = null,
-  //  @ColumnInfo(name = "book") val book: BookItem,
+    //  @ColumnInfo(name = "book") val book: BookItem,
     @ColumnInfo(name = "googleId") val googleId: String,
     @ColumnInfo(name = "kind") val kind: String?,
     @ColumnInfo(name = "title") val title: String?,
     @ColumnInfo(name = "subtitle") val subtitle: String?,
     @ColumnInfo(name = "year") val year: String?,
     @ColumnInfo(name = "averageRating") val averageRating: Double?,
-    @ColumnInfo(name = "personalRating") val personalRating: Int,
+    @ColumnInfo(name = "personalRating") val personalRating: Long,
     @ColumnInfo(name = "description") val description: String?,
     @ColumnInfo(name = "readDate") val readDate: String?,
     @ColumnInfo(name = "comment") val comment: String?
-  //  @ColumnInfo(name = "status") val status:Status?
+    //  @ColumnInfo(name = "status") val status:Status?
 )
 
 enum class BookStatus {
